+++ test.py
@@ -154,10 +154,11 @@ if __name__ == "__main__":
     assert_equality(WorkSheet, path.join(tests_path, 'test_wks_all.kicad_wks'))
     assert_equality(DesignRules, path.join(tests_path, 'test_dru_all.kicad_dru'))
     assert_equality(Board,     path.join(tests_path, 'test_arcs.kicad_pcb'))
+    assert_equality(Footprint, path.join(tests_path, 'test_fp_pad_newlines.kicad_mod'))
 
     print("\n---------------------\n")
 
     if global_passed:
-        print("✅ KiTools tests done, all green ✅")
+        print("✅ KiUtils tests done, all green ✅")
     else:
-        print("❌❌ KiTools tests done, but some tests failed! Check 'test.log' for more information ❌❌")
+        print("❌❌ KiUtils tests done, but some tests failed! Check 'test.log' for more information ❌❌")